requirements/common.txt
# Core Stuff
# -------------------------------------
Django==2.1.4

# Configuration
# -------------------------------------
argon2-cffi==18.3.0
django-environ==0.4.5
django-sites==0.10
python-dotenv==0.10.0
django-cors-headers==2.4.0



# Extensions
# -------------------------------------
pytz==2018.7

# Models
# -------------------------------------
psycopg2-binary==2.7.6.1

Pillow==5.3.0
django-extensions==2.1.4
django-uuid-upload-path==1.0.0
django-versatileimagefield==1.10

# REST APIs
# -------------------------------------
djangorestframework==3.9.0
django-rest-swagger==2.1.2

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Documentation
# -------------------------------------
mkdocs==1.0.4



# Auth Stuff
# -------------------------------------
PyJWT==1.7.1
django-mail-templated==2.6.5
